--- app/static/notation.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style_notation.css">
-    <title>Notation de film déjà vu</title>
-</head>
-    
-<body>
-    <div class="fenetre">
-        <div class="black-bar">
-            <header>
-                <nav>
-                    <ul class="nav-list">
-                        <li><a id="menu-btn"><img src="https://png.pngtree.com/png-vector/20190420/ourlarge/pngtree-menu-vector-icon-png-image_963342.jpg" alt="Bouton Image"></a></li>
-                        <li><a id="forYou" href="proposition.html">Pour toi</a></li>
-                        <li><a href="genres">Par genre</a></li>
-                        <li><a href="decouvrir">Découvrir</a></li>
-                    </ul>
-                </nav>
-            </header>
-        </div>
-        <h1>Noter un film déjà <span class="highlight">vu :</span> </h1>
-
-        <div class="search-container">
-            <input type="text" id="search-input" placeholder="Rechercher un film...">
-            <ul id="suggestions"></ul>
-            <select id="note">
-                <option value="0.5">0.5</option>
-                <option value="1.0">1.0</option>
-                <option value="1.5">1.5</option>
-                <option value="2.0">2.0</option>
-                <option value="2.5">2.5</option>
-                <option value="3.0">3.0</option>
-                <option value="3.5">3.5</option>
-                <option value="4.0">4.0</option>
-                <option value="4.5">4.5</option>
-                <option value="5.0">5.0</option>
-            </select>
-            
-        </div>
-        <button id="add">Ajouter</button>
-        
-    </div>
-    <div class="menu" id="menu">
-        <ul>
-            <li><a href="voir_apres.html">Liste des films à voir plus tard</a></li>
-            <li><a href="notation.html">Noter des nouveaux films</a></li>
-        </ul>
-        
-    </div>
-
-    <script src="notation_js.js"></script>
-</body>
-</html>
