+++ app/templates/proposition.html
@@ -1,0 +1,110 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{{ url_for('static', filename='style_proposition.css') }}">
+    <title>Proposition</title>
+</head>
+    
+<body>
+    
+    
+    <div class="fenetre">
+        <div class="black-bar">
+            <header>
+                <nav>
+                    <ul class="nav-list">
+                        <li><a id="menu-btn"><img src="https://png.pngtree.com/png-vector/20190420/ourlarge/pngtree-menu-vector-icon-png-image_963342.jpg" alt="Bouton Image"></a></li>
+                        <li><a href="{{ url_for('proposition') }}">Pour toi</a></li>
+                        <li><a href="genre">Par genre</a></li>
+                        <li><a href="decouvrir">Découvrir</a></li>
+                    </ul>
+                </nav>
+            </header>
+        </div>
+
+        <div class="position">
+            <h1 id="deja">Si déjà vu</h1>
+            
+        <div class="rating" id="rating">
+            <input type="radio" id="star10" name="rating" value="10" titre="10">
+            <label for="star10">&#9733;</label>
+            <input type="radio" id="star9" name="rating" value="9" titre="9">
+            <label for="star9">&#9733;</label>
+            <input type="radio" id="star8" name="rating" value="8" titre="8">
+            <label for="star8">&#9733;</label>
+            <input type="radio" id="star7" name="rating" value="7" titre="7">
+            <label for="star7">&#9733;</label>
+            <input type="radio" id="star6" name="rating" value="6" titre="6">
+            <label for="star6">&#9733;</label>
+            <input type="radio" id="star5" name="rating" value="5" titre="5">
+            <label for="star5">&#9733;</label>
+            <input type="radio" id="star4" name="rating" value="4" titre="4" >
+            <label for="star4">&#9733;</label>
+            <input type="radio" id="star3" name="rating" value="3" titre="3">
+            <label for="star3">&#9733;</label>
+            <input type="radio" id="star2" name="rating" value="2" titre="2">
+            <label for="star2">&#9733;</label>
+            <input type="radio" id="star1" name="rating" value="1" titre="1">
+            <label for="star1">&#9733;</label>   
+        </div>
+        
+        
+        <div class="info">
+            
+            <h2 id="resume">Résumé :</h2>
+            <p id="overview">Le paragraphe est une unité de contenu utilisée pour organiser des informations sur une page web. Il est généralement utilisé pour présenter du texte, mais peut également contenir d'autres types de contenu comme des images, des liens, des listes, etc.
+
+                Les paragraphes sont souvent utilisés pour diviser le contenu d'une page en sections distinctes, ce qui rend la lecture plus facile et agréable pour les utilisateurs. Ils permettent également aux développeurs web de structurer leur code de manière logique et compréhensible.
+                
+                En HTML, les paragraphes sont créés en utilisant la balise <code>&lt;p&gt;</code>. Cette balise est un élément de bloc, ce qui signifie qu'elle commence généralement sur une nouvelle ligne et occupe toute la largeur disponible de son conteneur parent.
+                
+                Voici un exemple de paragraphe HTMLdddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddjjjjjjjjjjjjjjjjjskwjekjfnjwfekjnwekjfnkjnwfefkjnwekjfnkjwefkjnddd bhjbhjbjhbhbhjbjhhjb</p>
+                    
+            <div id="acteurs">
+                <h2 id="acteur">Acteurs :</h2>
+                <p id="listacteur">Le texte que vous sossss s d a sd dsaa werq werwqeuhaitez écrire à côté de 'Acteurs'</p>
+            </div>
+            <div id="genres">
+                <h2 id="genre">Genres :</h2>
+                <p id="listegenre">Le texte que vous souhaitez écrire à côté de 'Acteurs'</p>
+            </div>
+            <div id="realisation_dates">
+                <h2 id="realisation_date">Date de publication :</h2>
+                <p id="date">Le texte que vous souhaitezdsdsddssdsddsdssdsds écrire à côté de 'Acteurs'</p>
+            </div>
+            <div id="realisateurs">
+                <h2 id="realisateur">Réalisateur :</h2>
+                <p id="crew">Le texte que vous souhaitez écrire à côté de 'Acteurs'</p>
+            </div>
+            <div id="boutons">
+                <button id="addListe">A voir plus tard</button>
+                <button id="passer">Passer</button>
+            </div>
+        </div>
+        <div class="image">
+            <img id="poster" src="https://global-img.gamergen.com/joker-2019-001_0900936868.jpg" alt="Image">
+        </div>
+        <div class="titre">
+            <h1 id="titre">Noter un film déjà </h1>
+        </div>
+       
+        
+    </div>
+
+    <div class="menu" id="menu">
+        <ul>
+            <li><a href="{{ url_for('voir_apres') }}">Liste des films à voir plus tard</a></li>
+            <li><a href="{{ url_for('notation') }}">Noter des nouveaux films</a></li>
+        </ul>
+        
+    </div>
+    
+
+
+
+    <script type="module"  src="{{ url_for('static', filename='proposition_js.js') }}"></script>
+    
+</body>
+</html>
