+++ app/templates/voir_apres.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{{ url_for('static', filename='style_voir_apres.css') }}">
+    <title>A voir plus tard</title>
+</head>
+<body>
+
+    <div class="fenetre">
+        <div class="black-bar">
+            <header>
+                <nav>
+                    <ul class="nav-list">
+                        <li><a id="menu-btn"><img src="https://png.pngtree.com/png-vector/20190420/ourlarge/pngtree-menu-vector-icon-png-image_963342.jpg" alt="Bouton Image"></a></li>
+                        <li><a href="{{ url_for('proposition') }}">Pour toi</a></li>
+                        <li><a href="genres">Par genre</a></li>
+                        <li><a href="decouvrir">Découvrir</a></li>
+                    </ul>
+                </nav>
+            </header>
+        </div>
+
+        <div class="container" id="container"></div>
+    </div>
+    <div class="menu" id="menu">
+        <ul>
+            <li><a href="{{ url_for('voir_apres') }}">Liste des films à voir plus tard</a></li>
+            <li><a href="{{ url_for('notation') }}">Noter des nouveaux films</a></li>
+        </ul>
+        
+    </div>
+
+
+    <script  src="{{ url_for('static', filename='voir_apres_js.js') }}"></script>
+</body>
+</html>
